Misc updates.

+++ opt/jdgregson/jdgregson-jupyter-host/setup.sh
@@ -41,13 +41,15 @@ NEEDRESTART_MODE=i apt-get install --yes \
     g++ \
     make
 
-mkdir -p /etc/apt/keyrings
-curl -fsSL https://deb.nodesource.com/gpgkey/nodesource-repo.gpg.key | sudo gpg --dearmor -o /etc/apt/keyrings/nodesource.gpg
-NODE_MAJOR=20
-echo "deb [signed-by=/etc/apt/keyrings/nodesource.gpg] https://deb.nodesource.com/node_$NODE_MAJOR.x nodistro main" | sudo tee /etc/apt/sources.list.d/nodesource.list
-apt-get update
-NEEDRESTART_MODE=i apt-get install --yes \
-    nodejs
+if [ ! -f /etc/apt/keyrings/nodesource.gpg ]; then
+    mkdir -p /etc/apt/keyrings
+    curl -fsSL https://deb.nodesource.com/gpgkey/nodesource-repo.gpg.key | sudo gpg --dearmor -o /etc/apt/keyrings/nodesource.gpg
+    NODE_MAJOR=20
+    echo "deb [signed-by=/etc/apt/keyrings/nodesource.gpg] https://deb.nodesource.com/node_$NODE_MAJOR.x nodistro main" | sudo tee /etc/apt/sources.list.d/nodesource.list
+    apt-get update
+    NEEDRESTART_MODE=i apt-get install --yes \
+        nodejs
+fi
 
 echo "$nrconf{restart} = 'i';" >> /etc/needrestart/needrestart.conf
 echo "$nrconf{kernelhints} = 0;" >> /etc/needrestart/needrestart.conf
